Finish setting up routing for pages
Finished setting up routing for all pages. This involved including
Navigation and Footer components within each page. The navigation
component has a sidebar component which contains href to specific
pages. Everything is working accorrdingly and the app now needs to be
styled appropriately. Could make the sidebar translate from the side
when the menu button is clicked. Work on that next.